+++ test/GitRemoteTest.php
@@ -39,7 +39,7 @@ class GitRemoteTest extends TestCase {
    */
   public function testJsonSerialize() {
     it('should return a map with default values for a newly created instance', function() {
-      $map = (new GitRemote)->jsonSerialize();
+      $map = (new GitRemote(''))->jsonSerialize();
       expect(get_object_vars($map))->to->have->lengthOf(2);
       expect($map->name)->to->be->empty;
       expect($map->url)->to->be->null;
@@ -50,21 +50,6 @@ class GitRemoteTest extends TestCase {
       expect(get_object_vars($map))->to->have->lengthOf(2);
       expect($map->name)->to->equal('origin');
       expect($map->url)->to->equal('https://github.com/cedx/coveralls.php.git');
-    });
-  }
-
-  /**
-   * @test GitRemote::setUrl
-   */
-  public function testSetUrl() {
-    it('should return an instance of `UriInterface` for strings', function() {
-      $url = (new GitRemote)->setUrl('https://github.com/cedx/akismet.php')->getUrl();
-      expect($url)->to->be->instanceOf(UriInterface::class);
-      expect((string) $url)->to->equal('https://github.com/cedx/akismet.php');
-    });
-
-    it('should return a `null` reference for unsupported values', function() {
-      expect((new GitRemote)->setUrl(123)->getUrl())->to->be->null;
     });
   }